marketing_supertool_v4_updated/requirements.txt
streamlit==1.37.0
pydantic==2.8.2
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
openai==1.43.0
gspread==6.1.4
google-auth==2.34.0
scikit-learn==1.5.1
plotly==5.24.1
beautifulsoup4==4.12.3
lxml==5.3.0
xlsxwriter==3.2.0
